let's go deeper into the console.verse
