+++ src/components/accueil/Accueil.tsx
@@ -11,7 +11,6 @@ import { WorldMap } from 'src/components/WorldMap';
 import type { FilterArgs } from 'src/components/accueil/Filters';
 import { Filters } from 'src/components/accueil/Filters';
 import { Activities } from 'src/components/activities/List';
-import { ClassroomContext } from 'src/contexts/classroomContext';
 import { UserContext } from 'src/contexts/userContext';
 import { VillageContext } from 'src/contexts/villageContext';
 import { useActivities } from 'src/services/useActivities';
@@ -21,8 +20,7 @@ import { UserType } from 'types/user.type';
 
 export const Accueil = () => {
   const { village, selectedPhase, setSelectedPhase } = React.useContext(VillageContext);
-  const { user } = React.useContext(UserContext);
-  const { classroom } = React.useContext(ClassroomContext);
+  const { user, getClassroomAsFamily } = React.useContext(UserContext);
   const isMediatorOrFamily = user !== null && user.type === (UserType.MEDIATOR || UserType.ADMIN || UserType.SUPER_ADMIN || UserType.FAMILY);
   const filterCountries = React.useMemo(
     () =>
@@ -96,7 +94,9 @@ export const Accueil = () => {
 
   console.log(user);
   console.log(village);
-  console.log(classroom);
+  if (user && user.id) {
+    console.log(getClassroomAsFamily(user.id));
+  }
 
   return (
     <Base showSubHeader>
